requirements.txt
accelerate==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
blake3==1.0.4
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1736182456865/work
bs4==0.0.2
cachetools==5.5.1
cbor==1.0.0
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.1
dataclasses-json==0.6.7
datasets==3.3.1
decorator==5.1.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
einops==0.8.1
executing==2.2.0
faiss-gpu-cu12==1.10.0
fastapi==0.115.8
filelock==3.17.0
FlagEmbedding==1.3.4
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
gpt4all==2.8.2
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna @ file:///croot/idna_1714398848350/work
ijson==3.3.0
importlib_metadata==8.6.1
iniconfig==2.0.0
inscriptis==2.5.3
interegular==0.3.3
ipython==8.32.0
ir_datasets==0.5.9
jedi==0.19.2
jieba==0.42.1
Jinja2 @ file:///croot/jinja2_1737760107953/work
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-huggingface==0.1.2
langchain-nomic==0.1.4
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
lark==1.2.2
lm-format-enforcer==0.10.10
loguru==0.7.3
lxml==5.3.1
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.26.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
nomic==3.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
ollama==0.4.7
openai==1.63.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
rank-bm25==0.2.2
ray==2.42.1
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.0
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
trec-car-tools==2.6
triton==3.1.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
unlzw3==0.2.3
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
vllm==0.7.2
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
zstandard==0.23.0
